split tests across multiple files for each specific function
